chore: add file rename to  test code and application
