+++ functional-test/e2e/add-note.spec.ts
@@ -1,4 +1,4 @@
-import { test } from "../lib/stepsFactory";
+import { test } from "../lib/steps.factory";
 
 test("As a caseworker add note to a case", async ({ addNoteSteps }) => {
