+++ package.json
@@ -13,8 +13,8 @@
     "test:preview": "echo preview test infra is wip",
     "test:aat": "cd sscs-ccd-e2e-tests && yarn install && yarn test:fullfunctional",
     "test:functional": "ENVIRONMENT=${ENVIRONMENT} ./bin/run-functional-tests.sh",
-    "test:functional-headless": "yarn playwright install && npx playwright test --config functional-test/config/playwright.config.ts --project chromium --reporter=html",
-    "test:functional-head": "yarn playwright install && npx playwright test --config functional-test/config/playwright.config.ts --project chromium --headed --reporter=html",
+    "test:functional-headless": "yarn playwright install && npx playwright test --config functional-test/playwright.config.ts --project chromium --reporter=html",
+    "test:functional-head": "yarn playwright install && npx playwright test --config functional-test/playwright.config.ts --project chromium --headed --reporter=html",
     "start": "node index.js",
     "test:coverage": "echo test:coverage",
     "test:a11y": "echo test:a11y",
